+++ resolwe_bio/tests/processes/test_reads_filtering.py
@@ -232,6 +232,26 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFiles(cutadapt_paired, 'fastq2', ['cutadapt_paired_reverse_trimmed.fastq.gz'],
                          compression='gzip')
 
+    def test_cutadapt_custom(self):
+        reads_single = self.prepare_reads(['cutadapt single.fastq.gz', 'cutadapt_single1.fastq.gz'])
+        reads_paired = self.prepare_paired_reads(mate1=['cutadapt forward1.fastq.gz', 'cutadapt_forward2.fastq.gz'],
+                                                 mate2=['cutadapt_reverse.fastq.gz'])
+
+        inputs_single = {'reads': reads_single.id}
+        inputs_paired = {'reads': reads_paired.id}
+
+        cutadapt_single = self.run_process('cutadapt-custom-single', inputs_single)
+        cutadapt_paired = self.run_process('cutadapt-custom-paired', inputs_paired)
+
+        self.assertFiles(cutadapt_single, 'fastq', ['cutadapt_custom_single_trimmed.fastq.gz'],
+                         compression='gzip')
+
+        self.assertFiles(cutadapt_paired, 'fastq', ['cutadapt_custom_paired_forward_trimmed.fastq.gz'],
+                         compression='gzip')
+
+        self.assertFiles(cutadapt_paired, 'fastq2', ['cutadapt_custom_paired_reverse_trimmed.fastq.gz'],
+                         compression='gzip')
+
     def test_cutadapt_amplicon(self):
         inputs = {
             'src1': ['56GSID_1k_mate1.fastq.gz'],